requirements.txt
Flask
flask_sqlalchemy
gunicorn
python-docx
docxcompose
selenium
beautifulsoup4
